refactor: all search bar
- make all search bars more customizable, showing selected label is optional
- move SearchOrderCustomer and SearchOrderDetail to orders
